deployment/requirements.txt
fastapi
uvicorn
torch
diffusers
pillow
